Move team related functions to service layer
