+++ services/org/user.go
@@ -2,7 +2,7 @@
 // Copyright 2019 The Gitea Authors. All rights reserved.
 // SPDX-License-Identifier: MIT
 
-package models
+package org
 
 import (
 	"context"
